Path problem
Fix PATH persistency
Fix condition to check the existence of directory (check if conda is installed, if the environnement exist

+++ Makefile
@@ -7,19 +7,32 @@ CONDA_HOME = $(HOME)/miniconda3
 CONDA_BIN_DIR = $(CONDA_HOME)/bin
 CONDA = $(CONDA_BIN_DIR)/conda
 
-ENV_DIR = $(CONDA_HOME)/envs/$(ENV_NAME)
+ENVS_DIR = $(CONDA_HOME)/envs/
+ENV_DIR = $(ENVS_DIR)/$(ENV_NAME)
 ENV_BIN_DIR = $(ENV_DIR)/bin
 ENV_LIB_DIR = $(ENV_DIR)/lib
 PYTHON = $(ENV_BIN_DIR)/python
 PIP=$(ENV_BIN_DIR)/pip3
 
 setup-run: setup run
-setup: conda-install conda-env-create packages-install
+setup: add-path conda-install conda-env-create packages-install
+
+####################################################################################################
+################################ Add conda to PATH #################################################
+#||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||#
+
+ifeq (,$(shell echo $(PATH) | grep $(CONDA_BIN_DIR)))
+PATH:=$(CONDA_BIN_DIR):$(PATH)
+endif
+add-path:
+ifneq (,$(wildcard $(CONDA_HOME)))
+	conda init
+endif
 
 ####################################################################################################
 ################# Download and install conda (if it's not installed yet) ###########################
 #||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||#
-ifeq (,$(shell which conda)) # check if conda is installed
+
 ifeq ($(OS),Windows_NT)
 CONDASH:=Miniconda3-latest-Windows-x86_64.exe
 else
@@ -32,32 +45,28 @@ endif # Linux
 endif # not Windows
 CONDAURL:=https://repo.anaconda.com/miniconda/$(CONDASH)
 
-ifeq (,$(shell echo $(PATH) | grep $(CONDA_BIN_DIR)))
-PATH:=$(CONDA_BIN_DIR):$(PATH)
-endif
-
-conda-install: # If conda is not installed
+conda-install:
+ifneq (,$(wildcard $(CONDA_HOME))) # check if conda is installed
+	@echo ">>> miniconda3 is installed in $(CONDA_HOME)"
+else 
+	@echo ">>>$(CONDAURL)"
 	@echo ">>> Setting up miniconda3..."
 	@echo ">>> Downloading $(CONDAURL) : ... "
 	@wget "$(CONDAURL)" && \
 	bash "$(CONDASH)" -b -p $(HOME)/miniconda3 && \
 	rm -f "$(CONDASH)"
-
-else # Else it is installed
-conda-install:
-	@echo ">>> miniconda3 is installed in $(CONDA_HOME)"
-endif # not HAS_CONDA
-
+	conda init
+endif 
 
 ####################################################################################################
 #################### Check if the conda environment exist (create if not) ##########################
 #||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||#
-conda-env-create:
-ifeq (,$(shell conda env list | grep $(ENV_NAME))) # check if the conda environment exist
+conda-env-create: conda-install
+ifneq (,$(wildcard $(ENV_DIR))) # check if the conda environment exist
+	@echo ">>> $(ENV_NAME) conda environment exist"
+else
 	conda create -n $(ENV_NAME) python=$(PYTHON_VERSION) -y
 	conda install -n $(ENV_NAME) pip -y
-else
-	@echo ">>> $(ENV_NAME) conda environment exist"
 endif
 
 ####################################################################################################
@@ -72,8 +81,7 @@ TORCH_INSTALL:=conda install pytorch torchvision torchaudio cudatoolkit=11.3 -c 
 endif
 #-------------- /EDIT CONDITIONNAL INSTALLATION ----------
 
-
-packages-install:
+packages-install: conda-env-create
 	conda update -n base -c defaults conda -y
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
@@ -95,16 +103,16 @@ packages-install:
 #---GIT INSTALLATION
 
 # { freqtrade
+ifeq (,$(shell $(PIP) list | grep freqtrade))
 	git clone https://github.com/freqtrade/freqtrade
 	source freqtrade/setup.sh
 	$(PIP) install freqtrade
 	rm -rf freqtrade
+endif
 # }
 
 #---CONDITIONNAL INSTALLATION
 	$(TORCH_INSTALL)
-
-
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
 
 run:
